refactor: extract taproot related logic to taproot.ts file
