+++ src/merkle.d.ts
@@ -1,3 +1,2 @@
 /// <reference types="node" />
 export declare function fastMerkleRoot(values: Buffer[], digestFn: (b: Buffer) => Buffer): Buffer;
-export declare function computeMastRoot(scripts: any): Buffer;
